Further split Actions into Recoverable and UnRecoverable component variants.
